A `send` on a closed EndPoint should fail with `SendClosed`.
The tests assume that a call to `closeEndPoint` results in a failed
EndPoint. The network-transport API semantics do not make this clear.
network-transport-tcp does "fail" an EndPoint when `closeEndPoint` is
used.
Though it might make more sense for it to close normally.

This changes the tests to expect a `TransportError SendClosed _` when
attempting a `send` on a closed EndPoint.